chore: simplify flutter secret mechanism even further
- remove singleton and make the class var static
- combine jwt secret and encrypt dart files into security dart
